+++ haskell/src/Text/IMF.hs
@@ -9,3 +9,4 @@ import           Text.IMF.Header   as Export
 import           Text.IMF.Mailbox  as Export
 import           Text.IMF.Message  as Export
 import           Text.IMF.MsgId    as Export
+import           Text.IMF.Network  as Export